requirements.txt
Django==2.2.1
django_countries==5.3.3
django_rest_swagger==2.2.0
requests==2.21.0
django_filter==2.1.0
django_filters==0.2.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
Pillow==6.0.0
